# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/purchase_invoices:
    get:
      tags:
        - PurchaseInvoices
      summary: List PurchaseInvoices
      description: Get a paginated list of PurchaseInvoices with optional filtering and sorting
      operationId: listPurchaseInvoices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseInvoiceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseInvoices
      summary: Create PurchaseInvoice
      description: Create a new PurchaseInvoice
      operationId: createPurchaseInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseInvoiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseInvoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/purchase_invoice_lines:
    get:
      tags:
        - PurchaseInvoiceLines
      summary: List PurchaseInvoiceLines
      description: Get a paginated list of PurchaseInvoiceLines with optional filtering and sorting
      operationId: listPurchaseInvoiceLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseInvoiceLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseInvoiceLines
      summary: Create PurchaseInvoiceLine
      description: Create a new PurchaseInvoiceLine
      operationId: createPurchaseInvoiceLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseInvoiceLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseInvoiceLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PurchaseInvoice:
      type: object
      required:
        - id
        - invoice_number
        - company_id
        - supplier_id
        - status
        - posting_date
        - early_pay_discount_percent
        - currency
        - net_total
        - tax_total
        - withholding_total
        - grand_total
        - outstanding_amount
        - payable_account_id
        - posting_state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        invoice_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        source_po_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        posting_date:
          type: string
          format: date
        due_date:
          nullable: true
          type: string
          format: date
        payment_term_id:
          nullable: true
          type: string
          format: uuid
        early_pay_discount_percent:
          type: string
        early_pay_discount_deadline:
          nullable: true
          type: string
          format: date
        early_pay_discount_account_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        net_total:
          type: string
        tax_total:
          type: string
        withholding_total:
          type: string
        grand_total:
          type: string
        outstanding_amount:
          type: string
        payable_account_id:
          type: string
          format: uuid
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePurchaseInvoiceRequest:
      type: object
      required:
        - invoice_number
        - company_id
        - supplier_id
        - status
        - posting_date
        - early_pay_discount_percent
        - currency
        - net_total
        - tax_total
        - withholding_total
        - grand_total
        - outstanding_amount
        - payable_account_id
        - posting_state
        - metadata
      properties:
        invoice_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        source_po_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        posting_date:
          type: string
          format: date
        due_date:
          nullable: true
          type: string
          format: date
        payment_term_id:
          nullable: true
          type: string
          format: uuid
        early_pay_discount_percent:
          type: string
        early_pay_discount_deadline:
          nullable: true
          type: string
          format: date
        early_pay_discount_account_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        net_total:
          type: string
        tax_total:
          type: string
        withholding_total:
          type: string
        grand_total:
          type: string
        outstanding_amount:
          type: string
        payable_account_id:
          type: string
          format: uuid
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PurchaseInvoiceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseInvoice'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PurchaseInvoiceLine:
      type: object
      required:
        - id
        - invoice_id
        - company_id
        - item_id
        - expense_account_id
        - quantity
        - unit_price
        - net_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        invoice_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        expense_account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        net_amount:
          type: string
        metadata:
          type: object
...
    CreatePurchaseInvoiceLineRequest:
      type: object
      required:
        - invoice_id
        - company_id
        - item_id
        - expense_account_id
        - quantity
        - unit_price
        - net_amount
        - metadata
      properties:
        invoice_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        expense_account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        net_amount:
          type: string
        metadata:
          type: object
...
    PurchaseInvoiceLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseInvoiceLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []